utils.py fn names:
More uniform naming schema
Changed names across various files
Separated db queries in the file
updated readme
Restructured files: data contains only db backups and variables (headers.py) that we draw on in the rest of the code. Utils now holds all utility functions
