+++ T2/jobvite_scraper.py
@@ -2,7 +2,7 @@ from curses.ascii import isdigit
 import requests
 from bs4 import BeautifulSoup
 import re
-from utilities.headers import headers
+from database_utils.headers import headers
 from utilities.utils import KEYWORDS, insert_jobs
 
 BASE_URL = "https://jobs.jobvite.com"
